Add files via upload
Modified the hint button routine call.  The original code used an interrupt routine to handle the hint processing.  This is not ideal since the process included subroutine calls, blocking delays, serial communication, etc.  I found that operation of the hint, new game, and shutdown was not reliable.  Sometimes I could not get new game or shutdown to work at all.
This update still maintains the hint button as an interrupt call, but now all the interrupt does is set a value of a status variable depending on the desired operation (hint request, new game, or shutdown).  The processing for the hint is now handled in the main code, during the user's move turn.  To start a new game, hold down the hint button, then hold down the OK button, then release the hint button.  To shutdown, hold down the hint button, then hold down the H8 button, then release the hint button.  I find this now works reliably.